+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased][unreleased]
 
+## [6.0.1] - 2020-11-19
+* Added `uniqueSlugStartIndex` test case
+* Fix `equal` -> `strictEqual`
+* Updated dependencies -> `found 0 vulnerabilities`
+
 ## [6.0.0] - 2020-09-29
 * Allow to configure unique slug start index, and make it 1 instead of 2
   to mimic what markdown-toc, github-slugger, and GitHub itself does by